% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{choose_nSignatures}
\alias{choose_nSignatures}
\title{Choose optimal number of signatures}
\usage{
choose_nSignatures(sample_by_component, nTry = 12, nrun = 50,
  cores = 1, seed = 77777, plot = TRUE)
}
\arguments{
\item{sample_by_component}{a sample-by-component \code{matrix}, generate from \code{generate_sbcMatrix} function.}

\item{nTry}{the maximal tried number of signatures, default is 12. Of note, this value should far less than number
of features or samples.}

\item{nrun}{a numeric giving the number of run to perform for each value in range of 2 to \code{nTry}, default is 50.
According to \code{NMF} package documentation, nrun set to 50 is enough to achieve robust result.}

\item{cores}{number of compute cores to run this task.}

\item{seed}{seed number.}

\item{plot}{\code{logical}. If \code{TRUE}, plot best rank survey.}
}
\value{
a \code{list} contains information of NMF run and rank survey.
}
\description{
This function use \code{NMF} package to evaluate the optimal number of signatures. The most
}
\examples{
\dontrun{
#' ## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "cnPattern"))
## generate copy-number features
tcga_features = derive_features(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = fit_mixModels(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = generate_sbcMatrix(tcga_features, tcga_components, cores = 1)
## optimal rank survey
 tcga_sig_choose = choose_nSignatures(tcga_sample_component_matrix,
 nrun = 10, cores = 1, plot = FALSE)
}

}
\author{
Geoffrey Macintyre, Shixiang Wang
}
