% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_main.R
\name{cnv_chooseSigNumber}
\alias{cnv_chooseSigNumber}
\title{Choose optimal number of signatures}
\usage{
cnv_chooseSigNumber(sample_by_component, nTry = 12, nrun = 10,
  cores = 1, seed = 77777, plot = TRUE, testRandom = TRUE)
}
\arguments{
\item{sample_by_component}{a sample-by-component \code{matrix}, generate from \code{cnv_generateSbCMatrix} function.}

\item{nTry}{the maximal tried number of signatures, default is 12. Of note, this value should far less than number
of features or samples.}

\item{nrun}{a numeric giving the number of run to perform for each value in range of 2 to \code{nTry}, default is 10.
According to \code{NMF} package documentation, nrun set to 50 is enough to achieve robust result.}

\item{cores}{number of compute cores to run this task.}

\item{seed}{seed number.}

\item{plot}{\code{logical}. If \code{TRUE}, plot best rank survey.}

\item{testRandom}{if generate random data from input to test measurements. Default is \code{TRUE}.}
}
\value{
a \code{list} contains information of NMF run and rank survey.
}
\description{
This function use \code{NMF} package to evaluate the optimal number of signatures. The most
}
\examples{
\dontrun{
#' ## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "VSHunter"))
## generate copy-number features
tcga_features = cnv_derivefeatures(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = cnv_fitMixModels(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = cnv_generateSbCMatrix(tcga_features, tcga_components, cores = 1)
## optimal rank survey
 tcga_sig_choose = cnv_chooseSigNumber(tcga_sample_component_matrix,
 nrun = 10, cores = 1, plot = FALSE)
}

}
\author{
Geoffrey Macintyre, Shixiang Wang
}
