% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_main.R
\name{cnv_pipe}
\alias{cnv_pipe}
\title{CNV signature pipeline}
\usage{
cnv_pipe(CN_data, cores = 1, genome_build = c("hg19", "hg38"),
  de_novo = TRUE, reference_components = NULL, min_comp = 2,
  max_comp = 10, min_prior = 0.001, model_selection = "BIC",
  nrep = 1, niter = 1000, nTry = 12, nrun = 10, seed = 123456,
  plot_survey = TRUE, testRandom = TRUE, nmfalg = "brunet",
  tmp = FALSE)
}
\arguments{
\item{CN_data}{a \code{QDNAseqCopyNumbers} object or a list contains multiple \code{data.frame}s
each one \code{data.frame} stores copy-number profile for one sample with 'chromosome', 'start', 'end' and
'segVal' these four necessary columns. Of note, 'segVal' column shoule be absolute copy number values.}

\item{cores}{number of compute cores to run this task. You can use \code{detectCores} function to check how
many cores you can use.}

\item{genome_build}{genome build version, must be one of 'hg19' or 'hg38'.}

\item{de_novo}{default is \code{TRUE}. If set to \code{FALSE}, it will use reference components to
generate sample-by-component matrix and then extract signatures.}

\item{reference_components}{the object result from \code{cnv_fitMixModels}, default is \code{NULL}. When \code{de_novo}
is \code{FALSE} and this
argument is \code{NULL}, it will use reference components from Nature Genetics paper.}

\item{min_comp}{minimal number of components to fit, default is 2.}

\item{max_comp}{maximal number of components to fit, default is 10.}

\item{min_prior}{minimal prior value, default is 0.001. Details about custom setting please
refer to \code{flexmix} package.}

\item{model_selection}{model selection strategy, default is 'BIC'.Details about custom setting please
refer to \code{flexmix} package.}

\item{nrep}{number of run times for each value of component, keep only the solution with maximum likelihood.}

\item{niter}{maximal number of iteration to achive converge.}

\item{nTry}{the maximal tried number of signatures, default is 12. Of note, this value should far less than number
of features or samples.}

\item{nrun}{the number of run to perform for each value in range of 2 to \code{nTry}, default is 10.
According to \code{NMF} package documentation, nrun set to 50 is enough to achieve robust result.}

\item{seed}{seed number.}

\item{plot_survey}{\code{logical}. If \code{TRUE}, plot best rank survey.}

\item{testRandom}{if generate random data from input to test measurements. Default is \code{TRUE}.}

\item{nmfalg}{specification of the NMF algorithm.}

\item{tmp}{whether create a tmp directory to store temp result or not, default is \code{FALSE}.}
}
\value{
a \code{list} contains results of NMF best rank survey, run, signature matrix, exposure list etc..
}
\description{
this pipeline integrate multiple independent steps in \code{VSHunter}.
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "VSHunter"))
## run cnv signature pipeline
result = cnv_pipe(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
}
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
