% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cnv_plotMixComponents}
\alias{cnv_plotMixComponents}
\title{Plot mixture fit model components}
\usage{
cnv_plotMixComponents(features, components, ...)
}
\arguments{
\item{features}{a \code{list} generate from \code{cnv_derivefeatures} function.}

\item{components}{a \code{list} contain \code{flexmix} object of copy-number features, obtain this
from \code{cnv_fitMixModels} function or use pre-compiled components data which come from CNV signature paper
https://www.nature.com/articles/s41588-018-0179-8 (set this argument as \code{NULL}).}

\item{...}{other options pass to \code{\link[cowplot]{plot_grid}} function of \code{cowplot} package.}
}
\value{
a ggplot object
}
\description{
Plot mixture fit model components
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_autoCaptureSignatures}},
  \code{\link{cnv_chooseSigNumber}},
  \code{\link{cnv_derivefeatures}},
  \code{\link{cnv_extractSignatures}},
  \code{\link{cnv_fitMixModels}},
  \code{\link{cnv_generateSbCMatrix}},
  \code{\link{cnv_getLengthFraction}},
  \code{\link{cnv_pipe}},
  \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotSignatures}},
  \code{\link{cnv_quantifySigExposure}},
  \code{\link{cnv_readprofile}}
}
\concept{CNV analysis functions}
