% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{derive_features}
\alias{derive_features}
\title{Derive copy number feature distributions}
\usage{
derive_features(CN_data, cores = 1, genome_build = c("hg19", "hg38"))
}
\arguments{
\item{CN_data}{a \code{QDNAseqCopyNumbers} object or a list contains multiple \code{data.frame}s
each one \code{data.frame} stores copy-number profile for one sample with 'chromosome', 'start', 'end' and
'segVal' these four necessary columns. Of note, 'segVal' column shoule be absolute copy number values.}

\item{cores}{number of compute cores to run this task.}

\item{genome_build}{genome build version, must be one of 'hg19' or 'hg38'.}
}
\value{
a \code{list} contains six copy number feature distributions.
}
\description{
This function summarise each copy-number profile using a number of different
feature distributions: sigment size, breakpoint number (per ten megabase), change-point copy-number,
segment copy-number, breakpoint number(per chromosome arm), length of segments with oscilating
copy-number
}
\examples{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "cnPattern"))
## generate copy-number features
tcga_features = derive_features(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")

}
\author{
Geoffrey Macintyre, Shixiang Wang
}
