% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{extract_Signatures}
\alias{extract_Signatures}
\title{Extract signature based on specified rank value}
\usage{
extract_Signatures(sample_by_component, nsig, seed = 77777,
  nmfalg = "brunet", cores = 1)
}
\arguments{
\item{sample_by_component}{a sample-by-component \code{matrix}, generate from \code{generate_sbcMatrix} function.}

\item{nsig}{specification of the factorization rank.}

\item{seed}{seed number.}

\item{nmfalg}{specification of the NMF algorithm.}

\item{cores}{number of compute cores to run this task.}
}
\value{
a object of \code{NMF} run.
}
\description{
Extract signature based on specified rank value
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "cnPattern"))
## generate copy-number features
tcga_features = derive_features(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = fit_mixModels(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = generate_sbcMatrix(tcga_features, tcga_components, cores = 1)
## optimal rank survey
 tcga_sig_choose = choose_nSignatures(tcga_sample_component_matrix, nrun = 10,
 cores = 1, plot = FALSE)
 tcga_signatures = extract_Signatures(tcga_sample_component_matrix, nsig = 3, cores = 1)
}
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
