% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{fit_mixModels}
\alias{fit_mixModels}
\title{Fit optimal number of mixture model components}
\usage{
fit_mixModels(CN_features, seed = 77777, min_comp = 2, max_comp = 10,
  min_prior = 0.001, model_selection = "BIC", nrep = 1,
  niter = 1000, cores = 1, featsToFit = seq(1, 6))
}
\arguments{
\item{CN_features}{a \code{list} generate from \code{derive_features} function.}

\item{seed}{seed number.}

\item{min_comp}{minimal number of components to fit, default is 2.}

\item{max_comp}{maximal number of components to fit, default is 10.}

\item{min_prior}{minimal prior value, default is 0.001. Details about custom setting please
refer to \code{flexmix} package.}

\item{model_selection}{model selection strategy, default is 'BIC'.Details about custom setting please
refer to \code{flexmix} package.}

\item{nrep}{number of run times fro each value of component, keep only the solution with maximum likelihood.}

\item{niter}{maximal number of iteration to achive converge.}

\item{cores}{number of compute cores to run this task.}

\item{featsToFit}{integer vector used for task assignment in parallel computation. Do not change it.}
}
\value{
a \code{list} contain \code{flexmix} object of copy-number features.
}
\description{
Apply mixture modelling to breakdown each feature distribution into mixtures
of Gaussian or mixtures of Poison distributions using the \code{flexmix} package.
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "cnPattern"))
## generate copy-number features
tcga_features = derive_features(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = fit_mixModels(CN_features = tcga_features, cores = 1)
}
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
