% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{generate_sbcMatrix}
\alias{generate_sbcMatrix}
\title{Generate a sample-by-component matrix}
\usage{
generate_sbcMatrix(CN_features, all_components = NULL, cores = 1,
  rowIter = 1000)
}
\arguments{
\item{CN_features}{a \code{list} contains six copy number feature distributions, obtain this from
\code{derive_features} function.}

\item{all_components}{a \code{list} contain \code{flexmix} object of copy-number features, obtain this
from \code{fit_mixModels} function or use pre-compiled components data which come from CNV signature paper
https://www.nature.com/articles/s41588-018-0179-8 (set this argument as \code{NULL}).}

\item{cores}{number of compute cores to run this task.}

\item{rowIter}{step size of iteration for rows of ech CNV feature \code{data.frame}.}
}
\value{
a numeric sample-by-component \code{matrix}
}
\description{
This function generate a sample-by-component matrix representing the sum of
posterior probabilities of each copy-number event being assigned to each component.
}
\examples{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "cnPattern"))
## generate copy-number features
tcga_features = derive_features(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = fit_mixModels(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = generate_sbcMatrix(tcga_features, tcga_components, cores = 1)
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
