% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NatGen_features}
\alias{NatGen_features}
\title{copy number features of 91 samples from Nat.Gen paper}
\format{A list}
\source{
Generate from https://www.nature.com/articles/s41588-018-0179-8
}
\description{
copy number features of 91 samples from Nat.Gen paper
}
\examples{
data(NatGen_features)
}
