% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_main.R
\name{nmf_extract_group}
\alias{nmf_extract_group}
\title{Extract groups (sample clustering) from NMF run results}
\usage{
nmf_extract_group(nmfObj, type = "consensus", matchConseOrder = FALSE)
}
\arguments{
\item{nmfObj}{NMF run results}

\item{type}{a character string that indicates the type of cluster membership should be returned，
valid input is 'samples' or 'consensus'.}

\item{matchConseOrder}{if \code{TRUE}, oder samples as they are shown in consensus map.}
}
\description{
X = W x H
W is the feature matrix, H is the sample matrix
After NMF run, use this function to select import features and assign groups for the two matrix
}
\details{
This code is modified from \href{https://github.com/naikai/sake/blob/master/R/nmf_utils.R}{source}.
}
\examples{
\dontrun{
nmf_extract_group(nmf_res)
}
}
\keyword{NMF}
\keyword{groups}
