% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancan12_absolute}
\alias{pancan12_absolute}
\title{absolute copy number file profile of ABSOLUTE pancan paper}
\format{A data.frame}
\source{
Generate from Synapse, paper link https://www.nature.com/articles/nbt.2203
}
\description{
absolute copy number file profile of ABSOLUTE pancan paper
}
\examples{
data(pancan12_absolute)
}
