% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancan12_info}
\alias{pancan12_info}
\title{clinical information of samples in ABSOLUTE pancan paper}
\format{A data.frame}
\source{
Generate from Synapse, paper link https://www.nature.com/articles/nbt.2203
}
\description{
clinical information of samples in ABSOLUTE pancan paper
}
\examples{
data(pancan12_info)
}
