% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{quantify_Signatures}
\alias{quantify_Signatures}
\title{quantify exposure for samples using Linear Combination Decomposition (LCD)}
\usage{
quantify_Signatures(sample_by_component, component_by_signature = NULL)
}
\arguments{
\item{sample_by_component}{a sample-by-component \code{matrix}, generate from \code{generate_sbcMatrix} function.}

\item{component_by_signature}{a componet by signature matrix, default is \code{NULL},
it will use pre-compiled data from CNV signature paper
https://www.nature.com/articles/s41588-018-0179-8}
}
\value{
a \code{list} contains absolute/relative exposure.
}
\description{
quantify exposure for samples using Linear Combination Decomposition (LCD)
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "VSHunter"))
## generate copy-number features
tcga_features = derive_features(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = fit_mixModels(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = generate_sbcMatrix(tcga_features, tcga_components, cores = 1)
## optimal rank survey
 tcga_sig_choose = choose_nSignatures(tcga_sample_component_matrix,
 nrun = 10, cores = 1, plot = FALSE)
 tcga_signatures = extract_Signatures(tcga_sample_component_matrix, nsig = 3, cores = 1)
 w = NMF::basis(tcga_signatures) # signature matrix
 tcga_exposure = quantify_Signatures(sample_by_component =
 tcga_sample_component_matrix, component_by_signature = w)
}
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
