% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snv_main.R
\name{snv_extractSignatures}
\alias{snv_extractSignatures}
\title{Extract mutational signatures from trinucleotide context.}
\usage{
snv_extractSignatures(mat, n = NULL, nrun = 1, nTry = 6,
  nrun_Try = 10, plotBestFitRes = FALSE, parallel = NULL,
  pConstant = NULL)
}
\arguments{
\item{mat}{Input matrix of diemnsion nx96 generated by \code{\link{trinucleotideMatrix}}}

\item{n}{decompose matrix into n signatures. Default NULL. Tries to predict best value for \code{n} by running NMF on a range of values and chooses based on cophenetic correlation coefficient.}

\item{nrun}{number of runs to perform on user specified rank.}

\item{nTry}{tries upto this number of signatures before choosing best \code{n}. Default 6.}

\item{nrun_Try}{number of runs to perform when obtain a best-fit rank for NMF.}

\item{plotBestFitRes}{plots consensus heatmap for range of values tried. Default FALSE}

\item{parallel}{calls to .opt argument of \code{\link{nmf}}. e.g, 'P4' for using 4 cores. See note on \code{\link{nmf}} for MAC users.}

\item{pConstant}{A small positive value to add to the matrix. Use it ONLY if the functions throws an \code{non-conformable arrays} error}
}
\value{
a list with decomposed scaled signatures, signature contributions in each sample and a cosine similarity table against validated signatures.
}
\description{
Decompose a matrix of 96 substitution classes into \code{n} signatures.
}
\details{
This function decomposes a non-negative matrix into n signatures.
Extracted signatures are compared against 30 experimentally validated signatures by calculating cosine similarity. See http://cancer.sanger.ac.uk/cosmic/signatures for details.
}
\examples{
\dontrun{
laml.tnm <- snv_trinucleotideMatrix(maf = laml, ref_genome = 'BSgenome.Hsapiens.UCSC.hg19',
    prefix = 'chr', add = TRUE, useSyn = TRUE)
laml.sign <- snv_extractSignatures(mat = laml.tnm, plotBestFitRes = FALSE)
}
}
\seealso{
Other SNV analysis functions: \code{\link{snv_plotSignatures}},
  \code{\link{snv_readprofile}},
  \code{\link{snv_signatureEnrichment}},
  \code{\link{snv_trinucleotideMatrix}}
}
\concept{SNV analysis functions}
