% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VSHunter.R
\docType{package}
\name{VSHunter}
\alias{VSHunter}
\alias{VSHunter-package}
\title{VSHunter: A package for capturing variation signatures from genomic data.}
\description{
The VSHunter package provides two categories of important functions:
one for CNV signature analysis and the other for mutation signature analysis.
}
