% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_main.R
\name{cnv_autoCaptureSignatures}
\alias{cnv_autoCaptureSignatures}
\title{Auto-capture signature and coresponding exposure}
\usage{
cnv_autoCaptureSignatures(sample_by_component, nTry = 12, nrun = 10,
  cores = 1, seed = 123456, plot = TRUE, testRandom = TRUE)
}
\arguments{
\item{sample_by_component}{a sample-by-component \code{matrix}, generate from \code{cnv_generateSbCMatrix} function.}

\item{nTry}{the maximal tried number of signatures, default is 12. Of note, this value should far less than number
of features or samples.}

\item{nrun}{the number of run to perform for each value in range of 2 to \code{nTry}, default is 10.
According to \code{NMF} package documentation, nrun set to 50 is enough to achieve robust result.}

\item{cores}{number of compute cores to run this task. You can use \code{\link[parallel:detectCores]{parallel::detectCores()}} function to check how
many cores you can use. If you are using \code{\link[=cnv_pipe]{cnv_pipe()}} feature, please do not use maximal number of
cores in your computer, it may cause some unexpected problems.}

\item{seed}{seed number.}

\item{plot}{\code{logical}. If \code{TRUE}, plot best rank survey.}

\item{testRandom}{if generate random data from input to test measurements. Default is \code{TRUE}.}
}
\value{
a \code{list} contains results of NMF best rank survey, run, signature matrix, exposure list etc..
}
\description{
this is a wrapper of \code{cnv_chooseSigNumber}, \code{cnv_extractSignatures}
and \code{cnv_quantifySigExposure} these three functions.
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "VSHunter"))
## generate copy-number features
tcga_features = cnv_derivefeatures(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = cnv_fitMixModels(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = cnv_generateSbCMatrix(tcga_features, tcga_components, cores = 1)
## optimal rank survey13
tcga_results = cnv_autoCaptureSignatures(tcga_sample_component_matrix, nrun=10, cores = 1)
}
}
\seealso{
\code{\link[=cnv_plotSignatures]{cnv_plotSignatures()}} for plot signatures and their exposures (contributions).
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
