% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_main.R
\name{cnv_derivefeatures}
\alias{cnv_derivefeatures}
\title{Derive copy number feature distributions}
\usage{
cnv_derivefeatures(CN_data, cores = 1, genome_build = c("hg19",
  "hg38"))
}
\arguments{
\item{CN_data}{a \code{QDNAseqCopyNumbers} object or a \code{list} contains multiple \code{data.frame}s (recommended),
each \code{data.frame} stores copy-number profile for one sample with 'chromosome', 'start', 'end' and
'segVal' these four necessary columns. Of note, 'segVal' column shoule be absolute copy number values.}

\item{cores}{number of compute cores to run this task.
You can use \code{\link[parallel:detectCores]{parallel::detectCores()}} function to check how
many cores you can use. If you are using \code{\link[=cnv_pipe]{cnv_pipe()}} feature,
please do not use maximal number of
cores in your computer, it may cause some unexpected problems.}

\item{genome_build}{genome build version, must be one of 'hg19' or 'hg38'.}
}
\value{
a \code{list} contains six copy number feature distributions.
}
\description{
This function summarise each copy-number profile using a number of different
feature distributions: sigment size, breakpoint number (per ten megabase), change-point copy-number,
segment copy-number, breakpoint number (per chromosome arm), length of segments with oscilating
copy-number.
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "VSHunter"))
## generate copy-number features
tcga_features = cnv_derivefeatures(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
}
}
\seealso{
\code{\link[=cnv_plotFeatureDistribution]{cnv_plotFeatureDistribution()}} for plotting feature distributions.

Other CNV analysis functions: \code{\link{cnv_autoCaptureSignatures}},
  \code{\link{cnv_chooseSigNumber}},
  \code{\link{cnv_extractSignatures}},
  \code{\link{cnv_fitMixModels}},
  \code{\link{cnv_generateSbCMatrix}},
  \code{\link{cnv_getLengthFraction}},
  \code{\link{cnv_pipe}},
  \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotMixComponents}},
  \code{\link{cnv_plotSignatures}},
  \code{\link{cnv_quantifySigExposure}},
  \code{\link{cnv_readprofile}}
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
\concept{CNV analysis functions}
