% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cnv_getLengthFraction}
\alias{cnv_getLengthFraction}
\title{Calculate length fraction profile of copy number}
\usage{
cnv_getLengthFraction(CN_data, genome_build = c("hg19", "hg38"),
  cols = c("Chromosome", "Start.bp", "End.bp", "modal_cn"),
  sample_col = "sample")
}
\arguments{
\item{CN_data}{a \code{QDNAseqCopyNumbers} object or a \code{list} contains multiple \code{data.frame}s (recommended),
each \code{data.frame} stores copy-number profile for one sample with 'chromosome', 'start', 'end' and
'segVal' these four necessary columns. Of note, 'segVal' column shoule be absolute copy number values.}

\item{genome_build}{genome build version, must be one of 'hg19' or 'hg38'.}

\item{cols}{four characters used to specify chromosome, start position,
end position and copy number value in input, respectively.
Default use names from ABSOLUTE calling result.}

\item{sample_col}{a character used to specify the sample column name.}
}
\value{
a data frame
}
\description{
Calculate length fraction profile of copy number
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_autoCaptureSignatures}},
  \code{\link{cnv_chooseSigNumber}},
  \code{\link{cnv_derivefeatures}},
  \code{\link{cnv_extractSignatures}},
  \code{\link{cnv_fitMixModels}},
  \code{\link{cnv_generateSbCMatrix}},
  \code{\link{cnv_pipe}},
  \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotMixComponents}},
  \code{\link{cnv_plotSignatures}},
  \code{\link{cnv_quantifySigExposure}},
  \code{\link{cnv_readprofile}}
}
\author{
Shixiang Wang \href{mailto:w_shixiang@163.com}{w_shixiang@163.com}
}
\concept{CNV analysis functions}
