% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_main.R
\name{cnv_quantifySigExposure}
\alias{cnv_quantifySigExposure}
\title{Quantify exposure for samples using Linear Combination Decomposition (LCD)}
\usage{
cnv_quantifySigExposure(sample_by_component,
  component_by_signature = NULL)
}
\arguments{
\item{sample_by_component}{a sample-by-component \code{matrix}, generate from \link{cnv_generateSbCMatrix} function.}

\item{component_by_signature}{a componet by signature matrix,
default is \code{NULL}, it will use pre-compiled data from CNV signature paper
https://www.nature.com/articles/s41588-018-0179-8.}
}
\value{
a \code{list} contains absolute/relative exposure.
}
\description{
Quantify exposure for samples using Linear Combination Decomposition (LCD)
}
\examples{
\dontrun{
## load example copy-number data from tcga
load(system.file("inst/extdata", "example_cn_list.RData", package = "VSHunter"))
## generate copy-number features
tcga_features = cnv_derivefeatures(CN_data = tcga_segTabs, cores = 1, genome_build = "hg19")
## fit mixture model  (this will take some time)
tcga_components = cnv_fitMixModels(CN_features = tcga_features, cores = 1)
## generate a sample-by-component matrix
tcga_sample_component_matrix = cnv_generateSbCMatrix(tcga_features, tcga_components, cores = 1)
## optimal rank survey
 tcga_sig_choose = cnv_chooseSigNumber(tcga_sample_component_matrix,
 nrun = 10, cores = 1, plot = FALSE)
 tcga_signatures = cnv_extractSignatures(tcga_sample_component_matrix, nsig = 3, cores = 1)
 w = NMF::basis(tcga_signatures) # signature matrix
 tcga_exposure = cnv_quantifySigExposure(sample_by_component =
 tcga_sample_component_matrix, component_by_signature = w)
}
}
\seealso{
Other CNV analysis functions: \code{\link{cnv_autoCaptureSignatures}},
  \code{\link{cnv_chooseSigNumber}},
  \code{\link{cnv_derivefeatures}},
  \code{\link{cnv_extractSignatures}},
  \code{\link{cnv_fitMixModels}},
  \code{\link{cnv_generateSbCMatrix}},
  \code{\link{cnv_getLengthFraction}},
  \code{\link{cnv_pipe}},
  \code{\link{cnv_plotDistributionProfile}},
  \code{\link{cnv_plotFeatureDistribution}},
  \code{\link{cnv_plotMixComponents}},
  \code{\link{cnv_plotSignatures}},
  \code{\link{cnv_readprofile}}
}
\author{
Geoffrey Macintyre, Shixiang Wang
}
\concept{CNV analysis functions}
