% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_main.R
\name{cnv_readprofile}
\alias{cnv_readprofile}
\title{Read copy number as a list of data.frame from data.frame or files}
\usage{
cnv_readprofile(input, is_dir = FALSE, pattern = NULL,
  ignore_case = FALSE, sep = "\\t", cols = c("Chromosome",
  "Start.bp", "End.bp", "modal_cn"), have_sampleCol = TRUE,
  sample_col = "sample")
}
\arguments{
\item{input}{a \code{data.frame} or a file or a directory contains copy number profile}

\item{is_dir}{is \code{input} a directory?}

\item{pattern}{a optional regular expression used to select part of files if input is a directory, more detail please see
\code{list.files} function.}

\item{ignore_case}{logical. Should pattern-matching be case-insensitive?}

\item{sep}{the field separator character. Values on each line of the file are separated by this character.}

\item{cols}{four characters used to specify chromosome, start position,
end position and copy number value, respectively. Default use names from ABSOLUTE calling result.}

\item{have_sampleCol}{Whether input have sample column or not.
This argument must be \code{TRUE} and \code{sample_col} also
properly be assigned when input is a file or a \code{data.frame}.}

\item{sample_col}{a character used to specify the sample column name.}
}
\value{
a \code{list} contains absolute copy-number profile for multiple samples.
}
\description{
This function is used to read copy number profile for preparing CNV signature
analysis.
}
\author{
Shixiang Wang <w_shixiang@163.com>
}
